Adjust alignment in DynamicCard component
The div container for the Image component in DynamicCard.tsx was removed. Instead, the 'items-start' css class is added to the parent div to align items at the start of the cross axis. Also, added "shrink-0" class to Image component to prevent shrinking. This was done to simplify the component's structure and improve image alignment for better UI/UX.